refactor(api): separate routes into accounts, node_settings
