+++ crates/interledger-api/src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod accounts;
+mod node_settings;
+
+pub use accounts::accounts_api;
+pub use node_settings::node_settings_api;
