+++ crates/interledger/Cargo.toml
@@ -43,7 +43,6 @@ config = "0.9.3"
 env_logger = "0.6.1"
 futures = "0.1.25"
 hex = "0.3.2"
-hyper = "0.12.28"
 interledger-api = { path = "../interledger-api", version = "^0.1.1-alpha.1", optional = true }
 interledger-btp = { path = "../interledger-btp", version = "^0.2.2-alpha.1", optional = true }
 interledger-ccp = { path = "../interledger-ccp", version = "^0.1.1-alpha.1", optional = true }
@@ -59,12 +58,10 @@ interledger-stream = { path = "../interledger-stream", version = "^0.2.2-alpha.1
 interledger-store-memory = { path = "../interledger-store-memory", version = "^0.2.2-alpha.1", optional = true }
 interledger-store-redis = { path = "../interledger-store-redis", version = "^0.2.2-alpha.1", optional = true}
 log = "0.4.6"
-parking_lot = "0.7.1"
 ring = "0.14.6"
 serde = "1.0.99"
 tokio = "0.1.20"
 url = "2.1.0"
-lazy_static = "1.3.0"
 libc = "0.2.62"
 
 [dev-dependencies]
